BUG: Fix tensorsolve for 0-sized input
`array.reshape(-1, size)` doesn't work with 0 in the dimensions.
The fix is to use explicit shape instead of `-1`. For "non-square"
tensors the `ValueError` would come from the reshape call, while previously
`LinAlgError` appeared from the solve call. To have the same error type
I added a check for squareness before the reshape.